refactored server-side code to cleaner packages
